move logic from parent to child (CommentAction, CommentForm)
