+++ src/components/Comment/Comment.jsx
@@ -8,73 +8,6 @@ import { AuthContext } from "../../context/AuthContext.jsx";
 
 const Comment = ({ data, showAuthModal }) => {
   const [showContent, setShowContent] = useState(false);
-  const [commentLikeDislike, setCommentLikeDislike] = useState({
-    ...data.commentLikeDislike,
-    first: true,
-  });
-  const auth = useContext(AuthContext);
-
-  const likeHandler = () => {
-    if (auth.isLoggedIn) {
-      setCommentLikeDislike((prev) => {
-        if (prev.ownStatus == "None") {
-          return {
-            dislikeCount: prev.dislikeCount,
-            likeCount: prev.likeCount + 1,
-            ownStatus: "Like",
-            first: false,
-          };
-        } else if (prev.ownStatus == "Like") {
-          return {
-            dislikeCount: prev.dislikeCount,
-            likeCount: prev.likeCount - 1,
-            ownStatus: "None",
-            first: false,
-          };
-        } else {
-          return {
-            dislikeCount: prev.dislikeCount - 1,
-            likeCount: prev.likeCount + 1,
-            ownStatus: "Like",
-            first: false,
-          };
-        }
-      });
-    } else {
-      showAuthModal();
-    }
-  };
-
-  const dislikeHandler = () => {
-    if (auth.isLoggedIn) {
-      setCommentLikeDislike((prev) => {
-        if (prev.ownStatus == "None") {
-          return {
-            dislikeCount: prev.dislikeCount + 1,
-            likeCount: prev.likeCount,
-            ownStatus: "Dislike",
-            first: false,
-          };
-        } else if (prev.ownStatus == "Dislike") {
-          return {
-            dislikeCount: prev.dislikeCount - 1,
-            likeCount: prev.likeCount,
-            ownStatus: "None",
-            first: false,
-          };
-        } else {
-          return {
-            dislikeCount: prev.dislikeCount + 1,
-            likeCount: prev.likeCount - 1,
-            ownStatus: "Dislike",
-            first: false,
-          };
-        }
-      });
-    } else {
-      showAuthModal()
-    }
-  };
 
   return (
     <div className={styles["comment"]}>
@@ -93,18 +26,17 @@ const Comment = ({ data, showAuthModal }) => {
         {(data.flag != "Spoiled" || showContent) && (
           <div className={styles["comment-action"]}>
             <CommentAction
-              data={commentLikeDislike}
-              likeHandler={likeHandler}
-              dislikeHandler={dislikeHandler}
+              data={data.commentLikeDislike}
+              showAuthModal={showAuthModal}
             />
           </div>
         )}
 
         {data.flag == "Spoiled" && showContent == false && (
-          <div className={styles["spoil"]}>
+          <div className={styles["spoil"]} onClick={() => setShowContent(true)}>
             <CircleWarningIcon />
             <span>این نظر حاوی اسپویلر است و داستان فیلم را لو می دهد.</span>
-            <span onClick={() => setShowContent(true)}>
+            <span>
               <ChevronDown />
             </span>
           </div>
