+++ src/components/Navbar/navbar.module.css
@@ -23,19 +23,19 @@
 }
 
 .nav-sm {
-  opacity: 0;
+  display: none;
   position: fixed;
   background-color: rgba(0, 0, 0, 0.7);
   top: 0;
   left: 0;
   height: 100vh;
   width: 100%;
-  z-index: -10;
+  z-index: -1;
   transition: all ease 0.3s;
 }
 
 .active {
-  opacity: 1;
+  display: block;
   z-index: 10;
 }
